requirements.txt
absl-py==2.1.0
adjustText==1.0.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.3.4
build==1.2.1
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
clusteval==2.2.1
clustimage==1.5.20
colorama==0.4.6
colorcet==3.0.1
coloredlogs==15.0.1
colourmap==1.1.16
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
datasets==2.19.2
datazets==0.1.9
debugpy==1.8.1
decorator==5.1.1
deepl==1.17.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distfit==1.7.3
distro==1.9.0
emoji==2.10.1
evaluate==0.4.2
executing==2.0.1
fastapi==0.110.2
fastjsonschema==2.19.0
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gpt4all==2.6.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
holoviews==1.18.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
hvplot==0.7.3
idna==3.6
ImageHash==4.3.1
imageio==2.34.0
imbalanced-learn==0.12.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
ismember==1.0.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaleido==0.2.1
keras==3.3.3
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.3
langchain-cli==0.0.24
langchain-community==0.2.4
langchain-core==0.2.5
langchain-experimental==0.0.60
langchain-huggingface==0.0.3
langchain-nvidia-ai-endpoints==0.1.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langgraph==0.0.69
langserve==0.2.1
langsmith==0.1.75
lazy_loader==0.3
libclang==18.1.1
libcst==1.4.0
linkify-it-py==2.0.3
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.33
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-postprocessor-flag-embedding-reranker==0.1.2
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
loguru==0.7.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nomic==3.0.23
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.33.0
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandas-dq==1.29
pandocfilters==1.5.0
panel==1.3.8
param==2.0.2
parso==0.8.4
patsy==0.5.6
pca==2.0.5
pillow==10.2.0
pinecone-client==3.2.2
platformdirs==4.2.2
plotly==5.19.0
posthog==3.5.0
prometheus-client==0.19.0
prompt_toolkit==3.0.46
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.5.0
pure-eval==0.2.2
py4j==0.10.9.7
pyamg==5.0.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyct==0.5.0
pydantic==2.7.0
pydantic_core==2.18.1
pydotplus==2.0.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.2.0
pypickle==1.1.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyviz_comms==3.0.1
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.32.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.3
scatterd==1.3.7
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
striprtf==0.0.26
sympy==1.12
tavily-python==0.3.3
tbb==2021.12.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
textblob==0.17.1
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
torch==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.9.4
types-python-dateutil==2.8.19.14
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.23.2
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
wordcloud==1.9.3
wrapt==1.16.0
xgboost==1.6.2
xlrd==2.0.1
xxhash==3.4.1
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.18.1
